+++ samples/Demo/Beef.Demo.Business/DataSvc/PersonDataSvc.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Beef.Demo.Business.DataSvc
+{
+    public static partial class PersonDataSvc
+    {
+        static PersonDataSvc()
+        {
+            _markOnAfterAsync = MarkOnAfterAsync;
+        }
+
+        private static async Task MarkOnAfterAsync()
+        {
+            await Beef.Events.Event.PublishAsync("Demo.Mark");
+        }
+    }
+}