+++ samples/Demo/Beef.Demo.Test/ReferenceDataTest.cs
@@ -3,6 +3,7 @@ using Beef.Demo.Common.Entities;
 using Beef.Test.NUnit;
 using Newtonsoft.Json.Linq;
 using NUnit.Framework;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 
@@ -92,6 +93,46 @@ namespace Beef.Demo.Test
                 {
                     x.Headers.IfNoneMatch.Add(new System.Net.Http.Headers.EntityTagHeaderValue(vals.First()));
                 }).GenderGetAllAsync());
+
+            AgentTester.Create<ReferenceDataAgent, GenderCollection>()
+                .ExpectStatusCode(HttpStatusCode.NotModified)
+                .Run((a) => a.Agent.GenderGetAllAsync(null, new WebApi.WebApiRequestOptions { ETag = vals.First() }));
+        }
+
+        [Test, Parallelizable]
+        public void A160_GetPowerSource_FilterByCodes()
+        {
+            var r = AgentTester.Create<ReferenceDataAgent, PowerSourceCollection>()
+                .ExpectStatusCode(HttpStatusCode.OK)
+                .Run((a) => a.Agent.PowerSourceGetAllAsync(new RefData.ReferenceDataFilter { Codes = new List<string> { "E", "n" } }));
+
+            Assert.IsNotNull(r);
+            Assert.IsNotNull(r.Value);
+            Assert.AreEqual(2, r.Value.Count());
+        }
+
+        [Test, Parallelizable]
+        public void A170_GetPowerSource_FilterByText()
+        {
+            var r = AgentTester.Create<ReferenceDataAgent, PowerSourceCollection>()
+                .ExpectStatusCode(HttpStatusCode.OK)
+                .Run((a) => a.Agent.PowerSourceGetAllAsync(new RefData.ReferenceDataFilter { Text = "el*" }));
+
+            Assert.IsNotNull(r);
+            Assert.IsNotNull(r.Value);
+            Assert.AreEqual(1, r.Value.Count());
+        }
+
+        [Test, Parallelizable]
+        public void A180_GetByCodes()
+        {
+            var r = AgentTester.Create<ReferenceDataAgent>()
+                .ExpectStatusCode(HttpStatusCode.OK)
+                .Run((a) => a.Agent.GetByCodesAsync(new WebApi.WebApiRequestOptions { UrlQueryString = "gender=m,f&powerSource=e&powerSource=f&eyecolor&$include=name,items.code" }));
+
+            Assert.IsNotNull(r);
+            Assert.IsNotNull(r.Content);
+            Assert.AreEqual("[{\"name\":\"Gender\",\"items\":[{\"code\":\"M\"},{\"code\":\"F\"}]},{\"name\":\"PowerSource\",\"items\":[{\"code\":\"E\"},{\"code\":\"F\"}]},{\"name\":\"EyeColor\",\"items\":[{\"code\":\"BLUE\"},{\"code\":\"BROWN\"},{\"code\":\"GREEN\"}]}]", r.Content);
         }
     }
 }