+++ src/Beef.Core/RefData/IReferenceDataCollection.cs
@@ -3,6 +3,7 @@
 using Beef.Entities;
 using System;
 using System.Collections;
+using System.Collections.Generic;
 
 namespace Beef.RefData
 {